Revert "ConnectionSettings"
This reverts commit ce9d6a38fd4b532cae1e0c80135c9992ddcb28bc.
