+++ rust/src/connection/credential.rs
@@ -17,17 +17,46 @@
  * under the License.
  */
 
-use std::fmt;
+use std::{fmt, fs, path::Path};
+
+use tonic::transport::{Certificate, ClientTlsConfig};
+
+use crate::Result;
 
 /// User credentials and TLS encryption settings for connecting to TypeDB Cloud.
 #[derive(Clone)]
 pub struct Credential {
     username: String,
     password: String,
+    is_tls_enabled: bool,
+    tls_config: Option<ClientTlsConfig>,
 }
 
 /// User credentials and TLS encryption settings for connecting to TypeDB Cloud.
 impl Credential {
+    /// Creates a credential with username and password. Specifies the connection must use TLS
+    ///
+    /// # Arguments
+    ///
+    /// * `username` --  The name of the user to connect as
+    /// * `password` -- The password for the user
+    /// * `tls_root_ca` -- Path to the CA certificate to use for authenticating server certificates.
+    ///
+    /// # Examples
+    ///
+    /// ```rust
+    /// Credential::with_tls(username, password, Some(&path_to_ca));
+    ///```
+    pub fn with_tls(username: &str, password: &str, tls_root_ca: Option<&Path>) -> Result<Self> {
+        let tls_config = Some(if let Some(tls_root_ca) = tls_root_ca {
+            ClientTlsConfig::new().ca_certificate(Certificate::from_pem(fs::read_to_string(tls_root_ca)?))
+        } else {
+            ClientTlsConfig::new()
+        });
+
+        Ok(Self { username: username.to_owned(), password: password.to_owned(), is_tls_enabled: true, tls_config })
+    }
+
     /// Creates a credential with username and password. The connection will not use TLS
     ///
     /// # Arguments
@@ -40,8 +69,8 @@ impl Credential {
     /// ```rust
     /// Credential::without_tls(username, password);
     ///```
-    pub fn new(username: &str, password: &str) -> Self {
-        Self { username: username.to_owned(), password: password.to_owned() }
+    pub fn without_tls(username: &str, password: &str) -> Self {
+        Self { username: username.to_owned(), password: password.to_owned(), is_tls_enabled: false, tls_config: None }
     }
 
     /// Retrieves the username used.
@@ -53,12 +82,23 @@ impl Credential {
     pub fn password(&self) -> &str {
         &self.password
     }
+
+    /// Retrieves whether TLS is enabled for the connection.
+    pub fn is_tls_enabled(&self) -> bool {
+        self.is_tls_enabled
+    }
+
+    pub fn tls_config(&self) -> &Option<ClientTlsConfig> {
+        &self.tls_config
+    }
 }
 
 impl fmt::Debug for Credential {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         f.debug_struct("Credential")
             .field("username", &self.username)
+            .field("is_tls_enabled", &self.is_tls_enabled)
+            .field("tls_config", &self.tls_config)
             .finish()
     }
 }
