+++ rust/tests/behaviour/steps/lib.rs
@@ -381,7 +381,7 @@ impl Context {
         _password: Option<&str>,
     ) -> TypeDBResult {
         assert!(!self.is_cloud);
-        let credential = Credential::new(_username.unwrap(), _password.unwrap());
+        let credential = Credential::without_tls(_username.unwrap(), _password.unwrap());
         let driver = TypeDBDriver::new_core(address, credential).await?;
         self.driver = Some(driver);
         Ok(())
