+++ rust/tests/integration/core/tmp.rs
@@ -5,7 +5,7 @@ use typedb_driver::{Credential, TypeDBDriver};
 #[serial]
 fn tmp() {
     async_std::task::block_on(async {
-        let driver: TypeDBDriver = TypeDBDriver::new_core(TypeDBDriver::DEFAULT_ADDRESS, Credential::new("admin", "password")).await.unwrap();
+        let driver: TypeDBDriver = TypeDBDriver::new_core(TypeDBDriver::DEFAULT_ADDRESS, Credential::without_tls("admin", "password")).await.unwrap();
         // let users = driver.users();
         // let x = users.contains("test").await;
         // users.create("user", "password").await.unwrap();
